requirements.txt
fastapi
facenet-pytorch
uvicorn
pillow
numpy
torch
python-multipart
chromadb
tqdm
uuid
typing